add ability to run tests in ephemeral namespaces and retain resources after failure

+++ common/src/main/java/io/quarkus/ts/openshift/common/OpenShiftTestExtension.java
@@ -15,9 +15,11 @@ import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ExtensionContext.Namespace;
 import org.junit.jupiter.api.extension.ExtensionContext.Store;
+import org.junit.jupiter.api.extension.LifecycleMethodExecutionExceptionHandler;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolutionException;
 import org.junit.jupiter.api.extension.ParameterResolver;
+import org.junit.jupiter.api.extension.TestExecutionExceptionHandler;
 import org.junit.jupiter.api.extension.TestInstancePostProcessor;
 
 import java.io.IOException;
@@ -31,7 +33,11 @@ import java.util.Optional;
 import static org.fusesource.jansi.Ansi.ansi;
 import static org.junit.platform.commons.util.AnnotationUtils.findAnnotatedFields;
 
-final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallback, BeforeEachCallback, TestInstancePostProcessor, ParameterResolver {
+// TODO at this point, this class is close to becoming unreadable, and could use some refactoring
+final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallback, BeforeEachCallback,
+        TestInstancePostProcessor, ParameterResolver,
+        LifecycleMethodExecutionExceptionHandler, TestExecutionExceptionHandler {
+
     private Store getStore(ExtensionContext context) {
         return context.getStore(Namespace.create(getClass()));
     }
@@ -63,12 +69,30 @@ final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallbac
         OpenShiftClient oc = getOpenShiftClient(context);
         AwaitUtil await = getAwaitUtil(context);
         return getStore(context)
-                .getOrComputeIfAbsent(OpenShiftUtil.class.getName(), ignogred -> new OpenShiftUtil(oc, await), OpenShiftUtil.class) ;
+                .getOrComputeIfAbsent(OpenShiftUtil.class.getName(), ignogred -> new OpenShiftUtil(oc, await), OpenShiftUtil.class);
     }
+
+    private void initTestsStatus(ExtensionContext context) {
+        getStore(context).put(TestsStatus.class.getName(), new TestsStatus());
+    }
+
+    private TestsStatus getTestsStatus(ExtensionContext context) {
+        TestsStatus testsStatus = getStore(context).get(TestsStatus.class.getName(), TestsStatus.class);
+        if (testsStatus == null) {
+            throw new IllegalStateException("missing " + TestsStatus.class.getSimpleName() + ", this is test framework bug");
+        }
+        return testsStatus;
+    }
+
+    // ---
 
     @Override
     public void beforeAll(ExtensionContext context) throws Exception {
         System.out.println("---------- OpenShiftTest set up ----------");
+
+        initTestsStatus(context);
+
+        createEphemeralNamespaceIfNecessary(context);
 
         AppMetadata metadata = getAppMetadata(context);
 
@@ -91,14 +115,32 @@ final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallbac
         getAwaitUtil(context).awaitAppRoute();
     }
 
+    private void createEphemeralNamespaceIfNecessary(ExtensionContext context) throws IOException, InterruptedException {
+        if (EphemeralNamespace.isEnabled()) {
+            EphemeralNamespace namespace = EphemeralNamespace.newWithRandomName();
+            getStore(context).put(EphemeralNamespace.class.getName(), namespace);
+
+            System.out.println(ansi().a("using ephemeral namespace ").fgYellow().a(namespace.name).reset());
+            new Command("oc new-project", "oc", "new-project", namespace.name).runAndWait();
+        }
+    }
+
     private void deployAdditionalResources(ExtensionContext context) throws IOException, InterruptedException {
+        TestsStatus testsStatus = getTestsStatus(context);
         Optional<AnnotatedElement> element = context.getElement();
         if (element.isPresent()) {
             AnnotatedElement annotatedElement = element.get();
             AdditionalResources[] annotations = annotatedElement.getAnnotationsByType(AdditionalResources.class);
             for (AdditionalResources additionalResources : annotations) {
-                AdditionalResourcesDeployed deployed = AdditionalResourcesDeployed.deploy(additionalResources);
-                getStore(context).put(new Object(), deployed);
+                AdditionalResourcesDeployed deployed = AdditionalResourcesDeployed.deploy(additionalResources, testsStatus);
+
+                if (EphemeralNamespace.isDisabled()) {
+                    // when using ephemeral namespaces, we don't delete additional resources because:
+                    // - when an ephemeral namespace is dropped, everything is destroyed anyway
+                    // - when retain on failure is enabled and failure occurs,
+                    //   everything in the ephemeral namespace must be kept intact
+                    getStore(context).put(new Object(), deployed);
+                }
             }
         }
     }
@@ -137,14 +179,51 @@ final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallbac
     public void afterAll(ExtensionContext context) throws Exception {
         System.out.println("---------- OpenShiftTest tear down ----------");
 
-        System.out.println("undeploying application");
-        new Command("oc delete", "oc", "delete", "-f", getResourcesYaml().toString(), "--ignore-not-found").runAndWait();
+        boolean shouldUndeployApplication = true;
+        if (EphemeralNamespace.isEnabled()) {
+            // when using ephemeral namespaces, we don't undeploy the application because:
+            // - when an ephemeral namespace is dropped, everything is destroyed anyway
+            // - when retain on failure is enabled and failure occurs,
+            //   everything in the ephemeral namespace must be kept intact
+            shouldUndeployApplication = false;
+        }
+        if (RetainOnFailure.isEnabled() && getTestsStatus(context).failed) {
+            shouldUndeployApplication = false;
+
+            if (EphemeralNamespace.isDisabled()) {
+                // when using ephemeral namespaces, a different message will be printed in dropEphemeralNamespaceIfNecessary
+                System.out.println(ansi().a("test ").fgYellow().a(context.getDisplayName()).reset()
+                        .a(" failed, not deleting any resources"));
+            }
+        }
+
+        if (shouldUndeployApplication) {
+            System.out.println("undeploying application");
+            new Command("oc delete", "oc", "delete", "-f", getResourcesYaml().toString(), "--ignore-not-found").runAndWait();
+        }
+
+        dropEphemeralNamespaceIfNecessary(context);
+    }
+
+    private void dropEphemeralNamespaceIfNecessary(ExtensionContext context) throws IOException, InterruptedException {
+        EphemeralNamespace ephemeralNamespace = getStore(context).get(EphemeralNamespace.class.getName(), EphemeralNamespace.class);
+        TestsStatus status = getTestsStatus(context);
+        if (ephemeralNamespace != null) {
+            if (RetainOnFailure.isEnabled() && status.failed) {
+                System.out.println(ansi().a("test ").fgYellow().a(context.getDisplayName()).reset()
+                        .a(" failed, keeping ephemeral namespace ").fgYellow().a(ephemeralNamespace.name).reset()
+                        .a(" intact"));
+            } else {
+                System.out.println(ansi().a("dropping ephemeral namespace ").fgYellow().a(ephemeralNamespace.name).reset());
+                new Command("oc delete", "oc", "delete", "project", ephemeralNamespace.name).runAndWait();
+            }
+        }
     }
 
     // ---
 
     @Override
-    public void beforeEach(ExtensionContext context) throws Exception {
+    public void beforeEach(ExtensionContext context) {
         System.out.println(ansi().a("---------- running test ")
                 .fgYellow().a(context.getParent().map(ctx -> ctx.getDisplayName() + ".").orElse(""))
                 .a(context.getDisplayName()).reset().a("----------"));
@@ -193,5 +272,41 @@ final class OpenShiftTestExtension implements BeforeAllCallback, AfterAllCallbac
             throw new OpenShiftTestException("Unsupported type " + injectionPoint.type().getSimpleName()
                     + " for @TestResource " + injectionPoint.description());
         }
+    }
+
+    // ---
+
+    @Override
+    public void handleBeforeAllMethodExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
+        failureOccured(context);
+        throw throwable;
+    }
+
+    @Override
+    public void handleBeforeEachMethodExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
+        failureOccured(context);
+        throw throwable;
+    }
+
+    @Override
+    public void handleAfterEachMethodExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
+        failureOccured(context);
+        throw throwable;
+    }
+
+    @Override
+    public void handleAfterAllMethodExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
+        failureOccured(context);
+        throw throwable;
+    }
+
+    @Override
+    public void handleTestExecutionException(ExtensionContext context, Throwable throwable) throws Throwable {
+        failureOccured(context);
+        throw throwable;
+    }
+
+    private void failureOccured(ExtensionContext context) {
+        getTestsStatus(context).failed = true;
     }
 }
